refactor: Cleaning up tsconfigs
