+++ .gitignore
@@ -53,7 +53,6 @@ test/fixtures/private
 **/*/tmp
 
 lib
-test-lib
 **/*/package-lock.json
 
 built
