--- index.ts
@@ -1,4 +1,0 @@
-/**
- * @since 1.1.0
- */
-export * from './lib/internal'
