+++ package.json
@@ -25,8 +25,8 @@
         "npm": ">= 8.x"
     },
     "license": "MIT",
-    "main": "lib/index.js",
-    "types": "lib/index.d.ts",
+    "main": "./index.js",
+    "typings": "./index.d.ts",
     "devDependencies": {
         "@types/chai": "^4.3.3",
         "@types/mocha": "9.1.1",
@@ -38,21 +38,21 @@
         "prettier": "^2.7.1",
         "tslint": "^6.1.3",
         "tslint-config-prettier": "^1.18.0",
+        "ts-node": "^10.9.1",
         "typescript": "^4.8.3"
     },
+    "dependencies": {
+    },
     "scripts": {
-        "clean": "rm -rf lib; rm -rf test-lib; rm -rf .test_run",
-        "test:compile": "tsc -b test",
+        "clean": "rm -rf lib; rm -rf .test_run",
         "test:run": "mocha --config .mocharc.js",
-        "test:build": "npm-run-all build test:compile",
-        "test": "npm-run-all test:build test:run",
+        "test": "npm run build && npm run test:run",
+        "compile": "tsc",
+        "build": "npm run clean && npm run compile",
         "report": "nyc report",
         "coverage": "nyc npm run test:run",
         "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
-        "lint": "tslint -p tsconfig.json",
-        "build:compile": "tsc -b src",
-        "build": "npm-run-all clean build:compile",
-        "prepare": "npm run build"
+        "lint": "tslint -p tsconfig.json"
     },
     "files": [
         "lib/**/*”"
