--- src/tsconfig.json
@@ -1,11 +1,0 @@
-{
-    "extends": "../tsconfig",
-    "include": [
-      "./**/*"
-    ],
-    "compilerOptions": {
-      "composite": true,
-      "rootDir": ".",
-      "outDir": "../lib"
-    }
-  }
